Небольшой рефакторинг
