+++ src/main/java/com/github/_1c_syntax/bsl/reader/common/ReaderUtils.java
@@ -1,0 +1,81 @@
+/*
+ * This file is a part of MDClasses.
+ *
+ * Copyright (c) 2019 - 2023
+ * Tymko Oleg <olegtymko@yandex.ru>, Maximov Valery <maximovvalery@gmail.com> and contributors
+ *
+ * SPDX-License-Identifier: LGPL-3.0-or-later
+ *
+ * MDClasses is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3.0 of the License, or (at your option) any later version.
+ *
+ * MDClasses is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with MDClasses.
+ */
+package com.github._1c_syntax.bsl.reader.common;
+
+import lombok.Getter;
+import lombok.experimental.UtilityClass;
+import org.apache.commons.io.FilenameUtils;
+
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.Arrays;
+
+/**
+ * Набор вспомогательных методов, используемых при чтении данных
+ */
+@UtilityClass
+public class ReaderUtils {
+  private static final byte[] PROTECTED_FILE_HEADER = new byte[]{-1, -1, -1, 127};
+
+  /**
+   * Читает информацию о "защищенности" модуля
+   *
+   * @param modulePath Путь к файлу модуля
+   * @return Параметы "защищенности"
+   */
+  public ProtectedModuleInfo readProtectedModuleInfo(Path modulePath) {
+    return new ProtectedModuleInfo(modulePath);
+  }
+
+  /**
+   * Описание "защищенности" модуля
+   */
+  public static class ProtectedModuleInfo {
+    @Getter
+    private boolean isProtected;
+    @Getter
+    private Path modulePath;
+
+    private ProtectedModuleInfo(Path path) {
+      modulePath = path;
+      isProtected = false;
+
+      if (!modulePath.toFile().exists()) {
+        var prtModulePath = Paths.get(FilenameUtils.removeExtension(modulePath.toFile().getPath()) + ".bin");
+        if (prtModulePath.toFile().exists()) {
+          isProtected = true;
+          modulePath = prtModulePath;
+        }
+      } else {
+        var bytes = new byte[PROTECTED_FILE_HEADER.length];
+        try (var fis = new FileInputStream(modulePath.toFile())) {
+          isProtected = (fis.read(bytes) == PROTECTED_FILE_HEADER.length
+            && Arrays.equals(bytes, PROTECTED_FILE_HEADER));
+        } catch (IOException e) {
+          // ошибка чтения в данном случае неважна
+        }
+      }
+    }
+  }
+}
